split simulation into submodule
